add 'monit_service' LWRP
Untested, but in theory you should just need to give it a name, start/stop 
command, and a 'script' (which is literal monit-dsl). It may be good to have
something where I can just have it import files I write from scratch, but that 
is a problem for another day.